# RecipeBook/docs/swagger.yaml
basePath: /
definitions:
  helpers.APIError:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
...
  helpers.DeleteResponse:
    properties:
      message:
        type: string
    type: object
...
  helpers.LikeInput:
    properties:
      recipe_id:
        type: integer
    type: object
...
  models.Comments:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      recipe:
        $ref: '#/definitions/models.Recipes'
      recipe_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
...
  models.Likes:
    properties:
      created_at:
        type: string
      id:
        type: integer
      recipe:
        $ref: '#/definitions/models.Recipes'
      recipe_id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
  models.Recipes:
    properties:
      category:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comments'
        type: array
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      ingredients:
        type: string
      likes:
        items:
          $ref: '#/definitions/models.Likes'
        type: array
      picture_url:
        type: string
      steps:
        type: string
      tags:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/models.Users'
      user_id:
        type: integer
    type: object
  models.Users:
    properties:
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      recipes:
        items:
          $ref: '#/definitions/models.Recipes'
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: soberkoder@swagger.io
    name: API Support
  description: This is a sample service for managing cars
  termsOfService: http://swagger.io/terms/
  title: RecipeBook API
  version: "1.0"
paths:
...
  /likes:
    get:
      consumes:
      - application/json
      description: Get details of all likes
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Likes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Get details
      tags:
      - Likes
    post:
      consumes:
      - application/json
      description: Post details of like corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: create like
        in: body
        name: helpers.LikeInput
        required: true
        schema:
          $ref: '#/definitions/helpers.LikeInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Likes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Post details for a given Id
      tags:
      - Likes
  /likes/{likeId}:
    delete:
      consumes:
      - application/json
      description: Delete the order corresponding to the input id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the like to be deleted
        in: path
        name: likeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.DeleteResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Delete like identified by the given id
      tags:
      - Likes
    get:
      consumes:
      - application/json
      description: Get details like corresponding to the input Id
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID of the like
        in: path
        name: likeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Likes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.APIError'
        "500":
          description: Server Error
          schema:
            $ref: '#/definitions/helpers.APIError'
      security:
      - BearerAuth: []
      summary: Get details for a given Id
      tags:
      - Likes
...
swagger: "2.0"